review: Move classes in statefun-smoke-e2e-driver to more specific packages
